requirements.txt
Flask==3.0.3
requests==2.32.3
pydub==0.25.1
PyYAML==6.0.1